+++ packages/spotlight/src/tests/utils-specs.js
@@ -1,4 +1,6 @@
 import {
+	getIntersectionRect,
+	getRect,
 	getPointRect,
 	getContainerRect,
 	isStandardFocusable
@@ -36,6 +38,81 @@ const scenarios = {
 };
 
 describe('utils', () => {
+	describe('#getIntersectionRect', () => {
+		test('should return the intersection rect between two given rects', () => {
+			const rect1 = {
+				getBoundingClientRect: () => ({
+					left: 0,
+					top: 0,
+					width: 100,
+					height: 100
+				})
+			};
+			const rect2 = {
+				getBoundingClientRect: () => ({
+					left: 50,
+					top: 50,
+					width: 100,
+					height: 100
+				})
+			};
+			const expected = {
+				left: 50,
+				top: 50,
+				width: 50,
+				height: 50,
+				right: 100,
+				bottom: 100,
+				center: {
+					x: 75,
+					y: 75,
+					left: 75,
+					right: 75,
+					top: 75,
+					bottom: 75
+				},
+				element: rect2
+			};
+			const actual = getIntersectionRect(rect1, rect2);
+
+			expect(actual).toEqual(expected);
+		});
+	});
+
+	describe('#getRect', () => {
+		test('should return the rect value calculated based on the given element', () => {
+			const element = {
+				getBoundingClientRect: () => ({
+					left: 10,
+					top: 20,
+					width: 100,
+					height: 200
+				})
+			};
+
+			const expected = {
+				left: 10,
+				top: 20,
+				width: 100,
+				height: 200,
+				right: 110,
+				bottom: 220,
+				center: {
+					x: 60,
+					y: 120,
+					left: 60,
+					right: 60,
+					top: 120,
+					bottom: 120
+				},
+				element
+			};
+			const actual = getRect(element);
+
+			expect(actual).toEqual(expected);
+		});
+	});
+
 	describe('#getPointRect', () => {
 		test('should return an rect value calculated based on given position', () => {
 			const expected = {